+++ packages/orq-rc/RELEASES.md
@@ -11209,3 +11209,13 @@ Based on:
 - [python v4.16.0-rc.19] packages/orq-rc
 ### Releases
 - [PyPI v4.16.0-rc.19] https://pypi.org/project/orq-ai-sdk/4.16.0-rc.19 - packages/orq-rc
+
+## 2026-09-25 14:36:58
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.799.0 (2.941.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.16.0-rc.20] packages/orq-rc
+### Releases
+- [PyPI v4.16.0-rc.20] https://pypi.org/project/orq-ai-sdk/4.16.0-rc.20 - packages/orq-rc